Add file via upload
This is a drawio file, I wish I can do it well but adding a new line or arrow is always in a wrong position/formatting. Always misaligns with the original ones.
For visualization: mainly about dynamic and interactive. I will add new methods too if I think up of some useful ways to implement it.